Move files to fix my awful rebase job
