+++ src/ObservationAssemblyRunner.cs
