+++ test/TestSpecification.cs
@@ -1,16 +1,35 @@
 using System;
-using System.Threading.Tasks;
+using Xunit;
+
+[assembly: TestFramework("Xunit.ObservationTestFramework", "Xunit.Bdd")]
 
 namespace Xunit.Bdd.Test
 {
 	public class behaves_like_a_specification : Specification
 	{
-		bool observedInBase = false;
+		private bool beforeObserveCalled = false;
+		private bool wasObservedAtBeforeObserve = true;
 
-		protected override Task ObserveAsync()
+		private bool observedInBase = false;
+
+		private bool afterObserveCalled = false;
+		private bool wasObservedAtAfterObserve = false;
+
+		protected override void BeforeObserve()
+		{
+			beforeObserveCalled = true;
+			wasObservedAtBeforeObserve = observedInBase;
+		}
+
+		protected override void Observe()
 		{
 			observedInBase = true;
-			return Task.CompletedTask;
+		}
+
+		protected override void AfterObserve()
+		{
+			afterObserveCalled = true;
+			wasObservedAtAfterObserve = observedInBase;
 		}
 
 		[Observation]
@@ -18,15 +37,35 @@ namespace Xunit.Bdd.Test
 		{
 			observedInBase.ShouldBeTrue("Observe should be called in the base class");
 		}
+
+		[Observation]
+		public void should_call_beforeobserve_before_observing()
+		{
+			beforeObserveCalled.ShouldBeTrue("BeforeObserve should have been called");
+			wasObservedAtBeforeObserve.ShouldBeFalse("BeforeObserve should have been called before Observe");
+		}
+
+		[Observation]
+		public void should_call_afterobserve_after_observing()
+		{
+			afterObserveCalled.ShouldBeTrue("AfterObserve should have been called");
+			wasObservedAtAfterObserve.ShouldBeTrue("AfterObserve should have been called after Observe");
+		}
+
+		[Observation(Skip="Skipped this observation")]
+		public void should_skip_this_observation()
+		{
+			Assert.True(false);
+		}
 	}
 
 	public class behaves_like_a_polymorphic_specification : behaves_like_a_specification
 	{
 		protected bool observedInDerived = false;
 
-		protected override async Task ObserveAsync()
+		protected override void Observe()
 		{
-			await base.ObserveAsync();
+			base.Observe();
 			observedInDerived = true;
 		}
 
@@ -40,9 +79,9 @@ namespace Xunit.Bdd.Test
 	public class TestException : Exception { }
 
 	[HandleExceptions]
-	public class behaves_like_a_specification_that_throws_when_observed : Specification
+	public class behaves_like_a_specification_that_throws_during_setup : Specification
 	{
-		protected override Task ObserveAsync()
+		protected override void Observe()
 		{
 			throw new TestException();
 		}
@@ -55,17 +94,23 @@ namespace Xunit.Bdd.Test
 		}
 	}
 
-	public class behaves_like_a_specification_that_unexpectedly_throws_when_observed : Specification
+	public class behaves_like_a_specification_that_unexpectedly_throws_during_setup : Specification
 	{
-		protected override Task ObserveAsync()
+		protected override void Observe()
 		{
 			throw new TestException();
 		}
 
-		[Observation(Skip = "This test should fail")]
-		public void should_fail()
+		[Observation]
+		public void should_be_inconclusive()
 		{
-			// This test will fail because of the exception thrown in Observe().
+			// This test will have an inconclusive result because of the exception thrown in Observe()
+		}
+
+		[Observation(Skip = "YOU SHOULD NEVER SEE THIS AS A TEST RESULT")] // The runner can't reach the point where it skips a test if its setup can't be run first.
+		public void should_still_be_inconclusive_even_if_skipped()
+		{
+			// This test will have an inconclusive result because of the exception thrown in Observe()
 		}
 	}
 }
